+++ src/ir_Whynter.cpp
@@ -22,8 +22,6 @@ void IRsend::sendWhynter(unsigned long data, int nbits) {
     // Set IR carrier frequency
     enableIROut(38);
 
-    noInterrupts();
-
     // Start
     mark(WHYNTER_BIT_MARK);
     space(WHYNTER_ZERO_SPACE);
@@ -35,8 +33,6 @@ void IRsend::sendWhynter(unsigned long data, int nbits) {
     // Data + stop bit
     sendPulseDistanceWidthData(WHYNTER_BIT_MARK, WHYNTER_ONE_SPACE, WHYNTER_BIT_MARK, WHYNTER_ZERO_SPACE, data, nbits, PROTOCOL_IS_MSB_FIRST,
     SEND_STOP_BIT);
-
-    interrupts();
 }
 
 //+=============================================================================